web_dev_python/price_of_chair/requirements.txt
request>=2.7.0
beautifulsoup4>=4.4.0